feat: add solution for esm.js
